feat(terminal): stack the telemetry into one scrollable left rail
Session and project navigation sat on the left while telemetry sat in a
rail of its own on the right, which split the window three ways and left
neither column enough width. The right rail is gone and its sections now
stack under the navigation lists in a single column that scrolls as one.

Every section publishes its own height from rail_geom, and the composer
advances by those same functions, so a section cannot draw somewhere the
layout did not account for. The painter and the hit-test both resolve
through nav_sections against the clamped offset, so a click lands on the
row that was drawn no matter how far the column has scrolled. A wheel
outside the rail rect falls through to the scrollback untouched.

Load, swap, disk and the throughput rates render as an em-dash: nothing
in the system measures them, and a plausible figure would be a lie.

+++ userland/capsule_terminal/src/paint/palette.rs
@@ -48,7 +48,7 @@ pub fn draw(fb: &mut PaintBuffer, body: Rect, pal: &Palette, ix: &Index, t: &The
 }
 
 fn draw_query(fb: &mut PaintBuffer, q: Rect, pal: &Palette, t: &Theme) {
-    let y = q.y + (q.h.saturating_sub(lh())) / 2;
+    let y = (q.y + (q.h.saturating_sub(lh())) / 2) as i32;
     let w = left(fb, q.x + RAIL_PAD, y, ">", t.accent);
     let text = core::str::from_utf8(pal.needle()).unwrap_or("");
     let x = q.x + RAIL_PAD + w + RAIL_GAP;
@@ -58,5 +58,5 @@ fn draw_query(fb: &mut PaintBuffer, q: Rect, pal: &Palette, t: &Theme) {
     }
     clipped(fb, x, y, q.x + q.w - x, text, t.fg);
     let cw = super::fit_text::width_of(fb, text, RAIL_PX);
-    fb.blend_rect(x + cw + 2, y, 2, lh(), t.accent);
+    fb.blend_rect(x + cw + 2, y as u32, 2, lh(), t.accent);
 }
